refactor(services): extract web agent memory state and client
Moves webAgentMemories and nextAgentMemoryId into
web-agent-memory-state.ts with the derivation, disambiguation and
injection helpers that are their write path. No binding is exported.

Three methods move into web-agent-memory-client.ts over the new
AgentMemoryService, whose signatures moved out of AgentService rather
than being copied.

The deferral note called this "3 methods over webAgentMemories /
nextAgentMemoryId", but five call sites read the pool: deleteApiAgent
counts memories for its delete-blocking check and sendMessage both
writes and indexes it. Both stay in the composition root and now go
through listWebAgentMemories, createWebAgentMemory and
simulateWebMemoryIndexInjection.

web-agent-client.ts 1608 -> 1514, agent-service.ts 346 -> 342.
Subtree aggregate 19164 -> 19203.

tsc --noEmit, contracts:check and the full suite pass: 286 files /
1301 tests, unchanged, with no test file edited.
